requirements.txt
altgraph==0.17.2
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
bcrypt==4.3.0
black==25.1.0
certifi==2025.1.31
cffi==1.15.1
click==8.1.8
colorama==0.4.6
databases==0.9.0
dotenv==0.9.9
ecdsa==0.19.1
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
isort==6.0.1
mypy-extensions==1.0.0
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
pefile==2024.8.26
platformdirs==4.3.6
pluggy==1.5.0
psycopg2==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyinstaller==5.2
pyinstaller-hooks-contrib==2025.1
pytest==8.3.5
python-dotenv==1.0.1
python-jose==3.4.0
pywin32-ctypes==0.2.3
rsa==4.9
ruff==0.9.10
setuptools==76.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
